I realized that I wasn't getting any rss from the Japanese rss links. There was an issue with the way that they made their xml list. The only solution I could really find was to use lxml, but with lxml I couldn't get everything else to work.
The solution I went with was to check if it was a Japanese site and then use lxml for it and I had to have a seperate method for that since the retrieval methods changed.

Ideally I would write a middleware that would interface with both of these, but I don't have too much time for that.